backend/requirements.txt
Flask
flask-cors
flask-bcrypt
flask-jwt-extended
python-dotenv
mysql-connector-python